HT-376: Improvements to LinkLabelHelper unit tests, handled another edge case. Minor code reorg.
Fixed bug in getting the name for the report file.
